Main API

DelaySSAdocs.DelayJumpSet — Type

mutable struct DelayJumpSet

A delay jump set that consists of five inputs at most, namely delay_trigger, delay_interrupt, delay_complete, delay_trigger_set and delay_interrupt_set.

Fields

delay_trigger

those reactions in the Markovian part that trigger the change of the state of the delay channels or/and the state of the reactants upon initiation.

delay_complete

those reactions in the Markovian part that change the state of the delay channels or/and the state of the reactants in the middle of on-going delay reactions.

delay_interrupt

those reactions that are initiated by delay trigger reactions and change the state of the delay channels or/and the state of the reactants upon completion.

delay_trigger_set

keys of delay_trigger.

delay_interrupt_set

keys of delay_interrupt.

Notes

delay_trigger::Dict{Int,Any}: those reactions in the Markovian part that trigger the change of the state of the delay channels or/and the state of the reactants upon initiation.

Keys: Indices of reactions defined in the Markovian part that can trigger the delay reactions;

Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

delay_interrupt::Dict{Int,Any}: those reactions in the Markovian part that change the state of the delay channels or/and the state of the reactants in the middle of on-going delay reactions.

Keys: Indices of reactions defined in the Markovian part that can interrupt the delay reactions;

Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

delay_complete::Dict{Int,Any}: those reactions that are initiated by delay trigger reactions and change the state of the delay channels or/and the state of the reactants upon completion.

Keys: Indices of the delay channel;

Values: Update functions(or Pair type) that decide how to update the delay channel or the state of the reactants.

delay_trigger_set::Vector{Int}: indices of reactions that can trigger the delay reaction.

delay_interrupt_set::Vector{Int}: indices of reactions that can interrupt the delay reactions.

We take this model for example.

delay_trigger_affect! = function (integrator, rng)
  append!(integrator.de_chan[1], τ)
end

# the 3rd reaction will trigger the delay reaction
delay_trigger = Dict(3=>delay_trigger_affect!)

# the 1st delay reaction will cause the 2nd species of molecule to degrade
delay_complete = Dict(1=>[2=>-1]) 
delay_affect! = function (integrator, rng)
   i = rand(rng, 1:length(integrator.de_chan[1]))
   deleteat!(integrator.de_chan[1],i)
end

# the 4th reaction will interrupt the delay reactions
delay_interrupt = Dict(4=>delay_affect!) 
delaysets = DelayJumpSet(delay_trigger,delay_complete,delay_interrupt)

DelaySSAdocs.DelayJumpProblem — Function

function DelayJumpProblem(prob::DiscreteProblem, aggregator::AbstractDelayAggregatorAlgorithm, jumps::JumpSet, delayjumpsets::DelayJumpSet, de_chan0)

Fields

prob::DiscreteProblem

A problem defined by some initial values in Markovian part.

aggregator::AbstractDelayAggregatorAlgorithm

The algorithm chose to solve the DelaySSA problem.

jumps::JumpSet

A jump set containing the information of Markovian part.

delayjumpsets::DelayJumpSet

A jump set containing the information of Non-Markovian part.

de_chan0::Vector{Vector{T}}

Initial condition of the delay channel and T::Float64.

djprob = DelayJumpProblem(dprob, DelayRejection(), jumpset, delayjumpset, de_chan0, save_positions=(true,true)).

function DelayJumpProblem(js::JumpSystem, prob, aggregator, delayjumpset, de_chan0; kwargs...)

Fields

js::JumpSystem

A jump system containing the information of Markovian part.

prob::DiscreteProblem

aggregator::AbstractDelayAggregatorAlgorithm

delayjumpsets::DelayJumpSet

de_chan0::Vector{Vector{T}}

djprob = DelayJumpProblem(jumpsys, dprob, DelayRejection(), delayjumpset, de_chan0, save_positions=(true,true))

Types and Algorithms

DelaySSAdocs.AbstractDelayAggregatorAlgorithm — Type

abstract type AbstractDelayAggregatorAlgorithm

An abstract type that contains delay stochastic simulation algorithms:

DelayDirect

DelaySSAdocs.DelayDirect — Type

struct DelayDirect <: DelaySSAdocs.AbstractDelayAggregatorAlgorithm

DelaySSAdocs.DelayRejection — Type

struct DelayRejection <: DelaySSAdocs.AbstractDelayAggregatorAlgorithm

DelaySSAdocs.DelayMNRM — Type

struct DelayMNRM <: DelaySSAdocs.AbstractDelayAggregatorAlgorithm

DelaySSAdocs.DelayDirectCR — Type

struct DelayDirectCR <: DelaySSAdocs.AbstractDelayAggregatorAlgorithm

A modifed Composition-Rejection Direct Method with delays (DelayDirectCR), implementation combining features from the original article and from the code in DiffEqJump package: DirectCR : A constant-time kinetic Monte Carlo algorithm for simulation of large biochemical reaction networks, by A. Slepoy, A.P. Thompson and S.J. Plimpton, J. Chem. Phys, 128, 205101 (2008). and Efficient Formulations for Exact Stochastic Simulation of Chemical Systems, by S. Mauch and M. Stalzer, ACM Trans. Comp. Biol. and Bioinf., 8, No. 1, 27-35 (2010).